backend/requirements.txt
asgiref==3.4.1
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
Django==3.2.5
django-cors-headers==3.7.0
django-rest-swagger==2.2.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
idna==2.10
itypes==1.2.0
Jinja2==3.0.1
jwt==1.2.0
MarkupSafe==2.0.1
openapi-codec==1.3.2
Pillow==8.3.1
pycparser==2.20
PyJWT==2.1.0
pytz==2021.1
requests==2.25.1
simplejson==3.17.3
sqlparse==0.4.1
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.6
whitenoise==5.2.0
